outsource definitons & set to service.options
